requirements.txt
google-generativeai>=0.8.0
pydantic>=2.10.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
httpx>=0.28.0
